requirements.txt
asyncpg==0.27.0
fastapi==0.95.1
starlette==0.26.1
uvicorn==0.22.0
pydantic==1.10.7
SQLAlchemy==2.0.12
python-dotenv==1.0.0
python-jose==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6